Make all build scripts indexed, cleanup unused plugin
