+++ build.gradle.kts
@@ -1,19 +1,25 @@
 import io.gitlab.arturbosch.detekt.Detekt
 import io.gitlab.arturbosch.detekt.report.ReportMergeTask
+import org.jetbrains.exposed.gradle.configureDetekt
+import org.jetbrains.exposed.gradle.configurePublishing
 import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
 
 plugins {
     kotlin("jvm") apply true
-    id("io.github.gradle-nexus.publish-plugin") apply true
     id("io.gitlab.arturbosch.detekt")
     id("org.jetbrains.kotlinx.binary-compatibility-validator") version "0.13.2"
 }
 
+repositories {
+    mavenLocal()
+    mavenCentral()
+}
+
 allprojects {
-    apply(from = rootProject.file("buildScripts/gradle/checkstyle.gradle.kts"))
+    configureDetekt()
 
     if (this.name != "exposed-tests" && this.name != "exposed-bom" && this != rootProject) {
-        apply(from = rootProject.file("buildScripts/gradle/publishing.gradle.kts"))
+        configurePublishing()
     }
 }
 
@@ -30,12 +36,14 @@ subprojects {
         detektPlugins("io.gitlab.arturbosch.detekt", "detekt-formatting", "1.21.0")
     }
     tasks.withType<Detekt>().configureEach detekt@{
-        enabled = this@subprojects.name !== "exposed-tests"
+        onlyIf { this@subprojects.name !== "exposed-tests" }
+
         finalizedBy(reportMerge)
         reportMerge.configure {
             input.from(this@detekt.xmlReportFile)
         }
     }
+
     tasks.withType<KotlinJvmCompile>().configureEach {
         kotlinOptions {
             jvmTarget = "1.8"
@@ -43,9 +51,4 @@ subprojects {
             languageVersion = "1.6"
         }
     }
-}
-
-repositories {
-    mavenLocal()
-    mavenCentral()
 }
